move all source to a separate package + update configs
