+++ ml_project/configs/estimator/log-reg.yaml
@@ -1,4 +1,4 @@
-_target_: settings.training_params.EstimatorConfig
+_target_: src.settings.params.EstimatorConfig
 
 model_artifact_path: ${dir_prefix}/artifacts/log-reg.pkl
 metrics_path: ${dir_prefix}/metrics/log-reg.json
