+++ ml_project/src/models/utils.py
@@ -14,7 +14,7 @@ from sklearn.metrics import (
     f1_score,
 )
 
-from settings.training_params import EstimatorConfig
+from ..settings.params import EstimatorConfig
 
 
 log = logging.getLogger(__name__)
