+++ ml_project/src/settings/params.py
@@ -1,0 +1,13 @@
+from .data_params import DatasetConfig, SplitConfig
+from .features_params import FeaturesConfig
+from .root_params import RootConfig, InfConfig
+from .training_params import EstimatorConfig
+
+__all__ = [
+    "DatasetConfig",
+    "SplitConfig",
+    "FeaturesConfig",
+    "RootConfig",
+    "InfConfig",
+    "EstimatorConfig"
+]
